fastapi_service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
pydantic==2.7.1
pydantic-settings==2.2.1

celery==5.4.0
redis==5.2.1

httpx==0.27.0

pytesseract==0.3.10
Pillow==12.1.0

aiosmtplib==2.0.1
jinja2==3.1.4
email-validator==2.1.0
python-multipart==0.0.9

pytest==8.2.1
pytest-asyncio==0.23.7
pytest-cov==5.0.0
pytest-mock==3.14.0

prometheus-client==0.20.0